# .junie/openapi.json
{
  "components": {
    "schemas": {
      "Body_login_login_post": {
        "properties": {
          "client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "client_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "format": "password",
            "title": "Client Secret"
          },
          "grant_type": {
            "anyOf": [
              {
                "pattern": "^password$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grant Type"
          },
          "password": {
            "format": "password",
            "title": "Password",
            "type": "string"
          },
          "scope": {
            "default": "",
            "title": "Scope",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "Body_login_login_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LoginData": {
        "description": "Data structure for login response",
        "properties": {
          "token_type": {
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "token_type"
        ],
        "title": "LoginData",
        "type": "object"
      },
      "LoginResponse": {
        "description": "Response schema for login endpoint",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/LoginData"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "message",
          "data"
        ],
        "title": "LoginResponse",
        "type": "object"
      },
...
      "MeResponse": {
        "description": "Response schema for me/profile endpoint",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/UserOut"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "message",
          "data"
        ],
        "title": "MeResponse",
        "type": "object"
      },
...
      "PasswordReset": {
        "properties": {
          "new_password": {
            "title": "New Password",
            "type": "string"
          },
          "token": {
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token",
          "new_password"
        ],
        "title": "PasswordReset",
        "type": "object"
      },
      "PasswordResetRequest": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "title": "PasswordResetRequest",
        "type": "object"
      },
      "PasswordResetResponse": {
        "description": "Response schema for password reset endpoints",
        "properties": {
          "data": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "message",
          "data"
        ],
        "title": "PasswordResetResponse",
        "type": "object"
      },
...
      "RefreshTokenData": {
        "description": "Data structure for refresh token response",
        "properties": {
          "token_type": {
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "token_type"
        ],
        "title": "RefreshTokenData",
        "type": "object"
      },
      "RefreshTokenResponse": {
        "description": "Response schema for refresh token endpoint",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RefreshTokenData"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "message",
          "data"
        ],
        "title": "RefreshTokenResponse",
        "type": "object"
      },
...
      "TagOut": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "value": {
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "id",
          "value"
        ],
        "title": "TagOut",
        "type": "object"
      },
...
      "UserOut": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "first_name": {
            "title": "First Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_superadmin": {
            "title": "Is Superadmin",
            "type": "boolean"
          },
          "last_name": {
            "title": "Last Name",
            "type": "string"
          },
          "tags": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/TagOut"
            },
            "title": "Tags",
            "type": "array"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "first_name",
          "last_name",
          "email",
          "is_superadmin"
        ],
        "title": "UserOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearerOrKey": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "contact": {
      "email": "votre.email@example.com",
      "name": "Votre Nom"
    },
    "description": "Backend FastAPI/SQLite (compatible PostgreSQL) pour la gestion des utilisateurs, environnements, groupes et permissions.\n\nA service wich deploys e home web platform on a Raspberry Pi or a VPS\nThis is usable for debian, Ubuntu systems. other Platforms are planned\n\n## API Overview\nThis API provides 108 operations across 78 endpoints. Supports DELETE, GET, POST, PUT HTTP methods.\n\n## Authentication\nMost endpoints (93%) require authentication.\n\n## Data Formats\nSupports the following content types: application/json, application/x-www-form-urlencoded.\n\n## Response Codes\nSuccess responses: 200. Error responses: 400, 401, 403, 404, 409, 418, 422, 500, 503.\n\n## Data Models\nIncludes 31 data model definitions for request and response validation.\n\n## Error Handling\nThe API returns standardized error responses with appropriate HTTP status codes and detailed error messages.",
    "license": {
      "name": "Apache",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "title": "App API",
    "version": "0.1.0",
    "x-generated-at": "2025-07-05T01:47:45.385528",
    "x-generator": "app OpenAPI Generator"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/login": {
      "post": {
        "description": "Authentifie un utilisateur avec son email et mot de passe et retourne un token d'accès",
        "operationId": "login_login_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_login_login_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "summary": "Successful response",
                    "value": {
                      "data": {},
                      "message": "Operation completed successfully",
                      "status": "success"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            },
            "description": "Login réussi"
          },
          "400": {
            "description": "Email ou mot de passe incorrect"
          },
          "422": {
            "content": {
              "application/json": {
                "examples": {
                  "error": {
                    "summary": "Client error",
                    "value": {
                      "detail": "Detailed error description",
                      "message": "Invalid request",
                      "status": "error"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Authentifier un utilisateur"
      }
    },
...
    "/me": {
      "get": {
        "description": "Récupère les informations du profil de l'utilisateur actuellement connecté",
        "operationId": "get_me_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "summary": "Successful response",
                    "value": {
                      "data": {},
                      "message": "Operation completed successfully",
                      "status": "success"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            },
            "description": "Information sur mon profil récupérée avec succès"
          },
          "401": {
            "description": "Non authentifié ou token invalide"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearerOrKey": []
          }
        ],
        "summary": "Profil utilisateur connecté"
      }
    },
...
    "/refresh-token": {
      "post": {
        "description": "Renouvelle le token d'accès à partir d'un refresh token valide",
        "operationId": "refresh_token_refresh_token_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "summary": "Successful response",
                    "value": {
                      "data": {},
                      "message": "Operation completed successfully",
                      "status": "success"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/RefreshTokenResponse"
                }
              }
            },
            "description": "Token renouvelé avec succès"
          },
          "400": {
            "description": "Refresh token invalide ou expiré, ou n'est pas un refresh token"
          }
        },
        "summary": "Renouveler le token"
      }
    },
...
    "/users/reset_password": {
      "post": {
        "description": "Réinitialise le mot de passe d'un utilisateur à l'aide d'un token de réinitialisation valide",
        "operationId": "reset_password_users_reset_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordReset"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "summary": "Successful response",
                    "value": {
                      "data": {},
                      "message": "Operation completed successfully",
                      "status": "success"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/PasswordResetResponse"
                }
              }
            },
            "description": "Mot de passe réinitialisé avec succès"
          },
          "400": {
            "description": "Token invalide, expiré ou incorrect"
          },
          "404": {
            "description": "Utilisateur non trouvé"
          },
          "422": {
            "content": {
              "application/json": {
                "examples": {
                  "error": {
                    "summary": "Client error",
                    "value": {
                      "detail": "Detailed error description",
                      "message": "Invalid request",
                      "status": "error"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Réinitialiser le mot de passe"
      }
    },
    "/users/reset_password_request": {
      "post": {
        "description": "Envoie un email contenant un lien de réinitialisation de mot de passe à l'adresse email fournie",
        "operationId": "reset_password_request_users_reset_password_request_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "summary": "Successful response",
                    "value": {
                      "data": {},
                      "message": "Operation completed successfully",
                      "status": "success"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/PasswordResetResponse"
                }
              }
            },
            "description": "Lien de réinitialisation envoyé par email (si l'email est enregistré)"
          },
          "422": {
            "content": {
              "application/json": {
                "examples": {
                  "error": {
                    "summary": "Client error",
                    "value": {
                      "detail": "Detailed error description",
                      "message": "Invalid request",
                      "status": "error"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Demander une réinitialisation de mot de passe"
      }
    },
...
  },
  "security": [
    {
      "OAuth2PasswordBearerOrKey": []
    }
  ],
  "servers": [
    {
      "description": "Development server (local)",
      "url": "http://localhost:8010"
    }
  ],
...
}